transform class updated
