+++ src/transform/usgs_transformer.py
@@ -2,28 +2,43 @@
 # Name: usgs_transformer.py
 # Author: Christopher O. Romanillos
 # Description: Class-based transformation of USGS API raw data 
-# Date: 09/02/25
+# Date: 09/18/25
 ##################################################################################
 
-from typing import List, Dict
+from typing import List, Dict, Any
 import pandas as pd
 from datetime import datetime
+
 from src.exceptions import TransformError
-from src.file_utils import DataManager
+
 
 class USGSTransformer:
-    def __init__(self, data_manager: DataManager, logger, endpoint_config: Dict):
-        self.data_manager = data_manager
-        self.logger = logger
-        self.endpoint_config = endpoint_config
+    REQUIRED_KEYS = {"data_manager", "logger", "endpoint_config"}
+
+    def __init__(self, **kwargs: Any):
+
+        # Validate required keys
+        missing = self.REQUIRED_KEYS - kwargs.keys()
+        if missing:
+            raise ValueError(f"Missing required arguments: {missing}")
+            
+        self.data_manager = kwargs["data_manager"]
+        self.logger = kwargs["logger"]
+        self.endpoint_config = kwargs["endpoint_config"]
+
+        # Optional attributes can still come through kwargs if needed later
+        self.__dict__.update(kwargs)
 
         # Map of built-in transformation functions
         self.transformation_funcs = {
             "yesno_to_bool": self.yesno_to_bool,
             "parse_datetime": self.parse_datetime,
-            "flatten_geometry": self.flatten_geometry
+            "flatten_geometry": self.flatten_geometry,
         }
 
+    ################################################
+    # Extraction Helpers
+    ################################################
     def extract_properties(self, records: List[Dict], include_geometry: bool = False) -> List[Dict]:
         extracted = []
         for feature in records:
@@ -33,9 +48,15 @@ class USGSTransformer:
             if include_geometry and feature.get("geometry") is not None:
                 props["geometry"] = feature["geometry"]
             extracted.append(props)
-        self.logger.info(f"Extracted {len(extracted)} property records (with id{' and geometry' if include_geometry else ''})")
+        self.logger.info(
+            f"Extracted {len(extracted)} property records "
+            f"(with id{' and geometry' if include_geometry else ''})"
+        )
         return extracted
 
+    ################################################
+    # Transformation Functions
+    ################################################
     def yesno_to_bool(self, value):
         if value is None:
             return None
@@ -71,10 +92,10 @@ class USGSTransformer:
             self.logger.warning(f"Failed to flatten geometry: {value}")
         return None
 
+    ################################################
+    # Normalization + DataFrame Conversion
+    ################################################
     def normalize_fields(self, records: List[Dict]) -> List[Dict]:
-        """
-        Apply endpoint-specific transformations as defined in config
-        """
         transformations = self.endpoint_config.get("transformations", {})
         for r in records:
             for field, func_name in transformations.items():
@@ -85,6 +106,9 @@ class USGSTransformer:
     def to_dataframe(self, records: List[Dict]) -> pd.DataFrame:
         return pd.DataFrame(records)
 
+    ################################################
+    # Full Transformation Workflow
+    ################################################
     def transform_latest_file(
         self,
         endpoint: str,
@@ -101,7 +125,12 @@ class USGSTransformer:
             normalized = self.normalize_fields(extracted)
 
             df = self.to_dataframe(normalized)
-            output_path = self.data_manager.save_dataframe(df, endpoint, use_processed=True, output_format=output_format)
+            output_path = self.data_manager.save_dataframe(
+                df,
+                endpoint,
+                use_processed=True,
+                output_format=output_format,
+            )
             self.logger.info(f"Transformed data saved to {output_path}")
             return str(output_path)
         except Exception as e:
